[FEATURE] Advanced metadata for File
The Core provides advanced metadata for File within a Core
extension "filemetadata". The extension is not activated
by default but can be when advanced metadata are needed.

The main benefit is to have unified metadata in TYPO3.
As a result, third-party extensions can build upon a common
base preventing segmentation.

Change-Id: I2e7ce7ac0d072f296b11dfa2189d32e5539f9994
Releases: 6.2
Resolves: #52719
Documentation: #52769
Reviewed-on: https://review.typo3.org/24715
Reviewed-by: Jigal van Hemert
Tested-by: Jigal van Hemert
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter