[BUGFIX] Category tab not shown for file when filemetadata installed
As a result of the TCA caching changes the order of
the manipulations done in the sys_file_metadata TCA
have changed. As a result the categories tab disappears
when you install ext:filemetadata.

This patch ads an extra check in the TCA override and
adds the categories tab when sys_file_metadata is categorized.

Resolves: #58620
Releases: 6.2
Change-Id: I75c7d7ccf40d5a2cc200c7d0a4e69f674024f628
Reviewed-on: https://review.typo3.org/29945
Reviewed-by: Jan Kiesewetter
Tested-by: Jan Kiesewetter
Reviewed-by: Oliver Hader
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel