[TASK] Group editable/read-only metadata fields
Instead of mixing editable and non-editable fields, it is best to
group them together.

In addition, "Duration" is one of the numerous metadata information
being related to either an audio or a video asset. Think ahead and
put the information on a dedicated tab, so that other extensions
know how to best extend support for metadata in FAL.

Change-Id: Iada3d47c7bad1e31981635d559489191dd6ca5ca
Resolves: #70949
Releases: master
Reviewed-on: https://review.typo3.org/44223
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>