[TASK] Increase size of `sys_file_metadata.publisher` database field
The database table `sys_file_metadata` set a limit of 45
characters for the `publisher` column (via TCA configuration), which
was not enough in some cases, for example:

  macOS Version 10.14.6 (Build 18G103) Quartz PDFContext

The character limit has now been raised to 120 characters.

The database column is not (by default) part of an index
column, and the table `sys_file_metadata` still has enough
column space available to rectify this change.

Resolves: #104872
Releases: main
Change-Id: I368654544f6d1cda0bfaf8880a86f87f6f658951
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86087
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>